ai_service/requirements.txt
# QCAA AI Grading Service
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.0
python-multipart>=0.0.9

# LLM
google-genai>=1.0.0

# Text extraction
pdfplumber>=0.11.0
python-docx>=1.1.0

# Embeddings / RAG
sentence-transformers>=3.0.0
torch>=2.0.0

# HTTP + config
requests>=2.32.0
python-dotenv>=1.1.0
